🚚 [승아] move: 컴포넌트 관련 파일 이동
